refactor: modularize codebase for improved maintainability
Moved configuration, logging, database, authentication, and common utility functions into dedicated modules.
