+++ src/models/certificate.py
@@ -1,15 +1,7 @@
 from pydantic import BaseModel, Field
 from datetime import date
 from typing import List
-
-class Signature(BaseModel):
-    id: str = Field(..., description="Unique ID of the signature document")
-    name: str = Field(..., description="Name of the signer")
-    post: str = Field(..., description="Position or designation of the signer")
-    image_b64: str = Field(..., description="Base64-encoded signature image")
-
-    class Config:
-        orm_mode = True
+from .signature import Signature
 
 class Certificate(BaseModel):
     id: str = Field(..., alias="_id", description="MongoDB ObjectId")
@@ -23,4 +15,4 @@ class Certificate(BaseModel):
     signatures: List[Signature] = Field(..., description="List of signature objects")
     
     class Config:
-        allow_population_by_field_name = True  
+        allow_population_by_field_name = True
