+++ src/utils/__init__.py
@@ -1,0 +1,20 @@
+from .logging_utils import setup_logging
+from .db_utils import (
+    setup_db,
+    seed_signatures,
+    seed_certificates,
+    seed_users,
+    get_certificate_by_credential,
+    get_signatures_by_ids,
+    get_user_by_email,
+    lifespan
+)
+from .auth_utils import (
+    authenticate_user,
+    create_access_token,
+    verify_password,
+    oauth2_scheme,
+    pwd_context,
+    process_login_request
+)
+from .common_utils import generate_credential_id
