+++ src/utils/logging_utils.py
@@ -1,0 +1,9 @@
+import logging
+
+def setup_logging(name: str = __name__) -> logging.Logger:
+    logging.basicConfig(
+        level=logging.INFO,
+        format="%(asctime)s [%(levelname)s] %(name)s: %(message)s",
+        datefmt="%Y-%m-%d %H:%M:%S"
+    )
+    return logging.getLogger(name)
